ISSUE #2653: Renamed UpsertAtheleteWeight to UpdateAthleteWeight
