+++ src/Controller/Api/V1/Settings/AthleteWeightRequestHandler.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App\Controller\Api\V1\Settings;
 
 use App\Domain\Athlete\Weight\DeleteAthleteWeight\DeleteAthleteWeight;
-use App\Domain\Athlete\Weight\UpsertAthleteWeight\UpsertAthleteWeight;
+use App\Domain\Athlete\Weight\UpdateAthleteWeight\UpdateAthleteWeight;
 use App\Domain\Settings\KeyValueBasedSettingsRepository;
 use App\Domain\Settings\SettingsRepository;
 use App\Infrastructure\CQRS\Command\Bus\CommandBus;
@@ -43,7 +43,7 @@ final readonly class AthleteWeightRequestHandler
     {
         $athleteWeightRequest = AthleteWeightRequest::fromRequest($request);
 
-        $this->commandBus->dispatch(UpsertAthleteWeight::from(
+        $this->commandBus->dispatch(UpdateAthleteWeight::from(
             on: $athleteWeightRequest->getOn(),
             weight: $athleteWeightRequest->getWeight(),
         ));
